% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antal_indlagt.R
\name{antal_indlagt}
\alias{antal_indlagt}
\title{Wrapper for C++ function of number admitted per time unit}
\usage{
antal_indlagt(admissions, time_units, debug = FALSE)
}
\arguments{
\item{within_hosp}{}
}
\description{
Wrapper for C++ function of number admitted per time unit
}
