% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiLPRNetwork-package.R
\docType{package}
\name{EpiLPRNetwork-package}
\alias{EpiLPRNetwork}
\alias{EpiLPRNetwork-package}
\title{EpiLPRNetwork: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

}
\keyword{internal}
