% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edges.R
\name{get_edges}
\alias{get_edges}
\title{Title}
\usage{
get_edges(
  all_by_mth = NA,
  year_group = TRUE,
  risk_function = NULL,
  output_level = "month",
  max_date = "2019-01-01 00:00",
  retry = Inf,
  timeout = c(1, 5, 15, 60),
  testing = FALSE
)
}
\description{
Title
}
