% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sql-table.R
\name{get_sql_table}
\alias{get_sql_table}
\title{Get an SQL connection to the specified table}
\usage{
get_sql_table(table_name)
}
\arguments{
\item{table_name}{the name of the table to return}
}
\description{
Get an SQL connection to the specified table
}
