% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_risk.R
\name{patient_risk}
\alias{patient_risk}
\title{Wrapper for C++ function of patient transfer risk}
\usage{
patient_risk(trisk, within_hosp = FALSE)
}
\arguments{
\item{within_hosp}{}
}
\description{
Wrapper for C++ function of patient transfer risk
}
